Move almost all code from Queries::global_ctxt into passes::create_global_ctxt
